Hamiltonian splitting

Hamiltonian splitting

GEMPIC.HamiltonianSplitting — Type.

HamiltonianSplitting( maxwell_solver,
                      kernel_smoother_0, kernel_smoother_1,
                      particle_group,
                      e_dofs_1, e_dofs_2, b_dofs,
                      x_min, Lx)

Hamiltonian splitting type for Vlasov-Maxwell

Integral over the spline function on each interval (order p+1)

Integral over the spline function on each interval (order p)

e_dofs describing the two components of the electric field

b_dofs describing the magnetic field

j_dofs for kernel representation of current density.

source

GEMPIC.operatorHB — Method.

operatorHB(h, dt)

Push HB: Equations to be solved ``Vnew = V_old``

\[\begin{eqnarray} \partial_t E_1 = 0 & -> & E_{1,new} = E_{1,old} \\ \partial_t E_2 = - \partial_{x_1} B & -> & E_{2,new} = E_{2,old}-dt*\partial_{x_1} B \\ \partial_t B = 0 & -> & B_new = B_old \\ \end{eqnarray}\]

source

GEMPIC.operatorHE — Method.

operatorHE(h, dt)

Push H_E: Equations to be solved

\[\begin{eqnarray} \partial_t f + E_1 \partial_{v_1} f + E_2 \partial_{v_2} f = 0 &->& V_new = V_old + dt * E \\ \partial_t E_1 = 0 &->& E_{1,new} = E_{1,old} \\ \partial_t E_2 = 0 &->& E_{2,new} = E_{2,old} \\ \partial_t B + \partial_{x_1} E_2 = 0 &->& B_{new} = B_{old} - dt \partial_{x_1} E_2 \end{eqnarray}\]

source

GEMPIC.operatorHp1 — Method.

operatorHp1(h, dt)

\[\begin{eqnarray} \partial_t f + v_1 \partial_{x_1} f = 0 & -> & X_{new} = X_{old} + dt V_1 \\ V_{new},2 = V_{old},2 + \int_0 h V_{old},1 B_{old} && \\ \partial_t E_1 = - \int v_1 f(t,x_1, v) dv & -> & E_{1,new} = E_{1,old} - \int \int v_1 f(t,x_1+s v_1,v) dv ds && \\ \partial_t E_2 = 0 & -> & E_{2,new} = E_{2,old} \\ \partial_t B = 0 & => & B_{new} = B_{old} \end{eqnarray}\]

Here we have to accumulate j and integrate over the time interval. At each $k=1,...,n_{grid}$, we have for $s \in [0,dt]: j_k(s) = \sum_{i=1,..,N_p} q_i N((x_k+sv_{1,k}-x_i)/h) v_k$, where $h$ is the grid spacing and $N$ the normalized B-spline In order to accumulate the integrated $j$, we normalize the values of $x$ to the grid spacing, calling them $y$, we have

\[ j_k(s) = \sum_{i=1,..,N_p} q_i N(y_k+\frac{s}{h} v_{1,k}-y_i) v_k.\]

Now, we want the integral

\[\int_{0..dt} j_k(s) d s = \sum_{i=1}^{N_p} q_i v_k \int_{0}{dt} N(y_k+\frac{s}{h} v_{1,k}-y_i) ds = \sum_{i=1}^{N_p} q_i v_k \int_{0}^{dt} N(y_k + w v_{1,k}-y_i) dw\]

For each particle compute the index of the first DoF on the grid it contributes to and its position (normalized to cell size one). Note: j_dofs does not hold the values for j itself but for the integrated j.

Then update particle position: $X_{new} = X_{old} + dt * V$

source

GEMPIC.operatorHp2 — Method.

operatorHp2(h, dt)

Push Hp2: Equations to solve are

\[\begin{eqnarray} X_{new} = X_{old} && \\ V_{new,1} = V_{old,1} + \int_0 h V_{old,2} B_{old} && \\ \partial_t E_1 = 0 & => & E_{1,new} = E_{1,old} \\ \partial_t E_2 = - \int v_2 f(t,x_1, v) dv & => & E_{2,new} = E_{2,old} - \int \int v_2 f(t,x_1 + s v_1,v) dv ds\\ \partial_t B = 0 & => & B_{new} = B_{old} && \\ \end{eqnarray}\]

source

GEMPIC.lie_splitting — Method.

lie_splitting( h, dt, number_steps)

Lie splitting

source

GEMPIC.lie_splitting_back — Method.

lie_splitting_back(h, dt, number_steps)

Lie splitting (oposite ordering)

source

GEMPIC.strang_splitting — Method.

strang_splitting( h, dt, number_steps)

Strang splitting

time splitting object

time step

number of time steps

source