Hamiltonian splitting Spin

GEMPIC.HamiltonianSplittingSpin — Type

HamiltonianSplittingSpin( maxwell_solver,
                          kernel_smoother_0, kernel_smoother_1,
                          particle_group, e_dofs, a_dofs )

Hamiltonian splitting type for Vlasov-Maxwell

Integral over the spline function on each interval (order p+1)

Integral over the spline function on each interval (order p)

e_dofs describing the two components of the electric field

a_dofs describing the potential vector

GEMPIC.operatorHA — Method

operatorHA(h, dt)

\[\begin{aligned} \dot{p} = (A_y, A_z) \cdot \partial_x (A_y, A_z) \\ \dot{Ey} = -\partial_x^2 A_y + A_y \rho \\ \dot{Ez} = -\partial_x^2 A_z + A_z \rho \\ \end{aligned}\]

GEMPIC.operatorHE — Method

operatorHE(h, dt)

\[\begin{aligned} \dot{v} & = E_x \\ \dot{A}_y & = -E_y \\ \dot{A}_z & = -E_z \end{aligned}\]

GEMPIC.operatorHp — Method

operatorHp(h, dt)

\[\begin{aligned} \dot{x} & =p \\ \dot{E}_x & = - \int (p f ) dp ds \end{aligned}\]

GEMPIC.operatorHs — Method

operatorHs(h, dt)

Push H_s: Equations to be solved

\[\begin{aligned} \dot{s} &= s x B = (s_y \partial_x A_y +s_z \partial_x Az, -s_x \partial_x A_y, -s_x \partial_x A_z) \\ \dot{p} &= s \cdot \partial_x B = -s_y \partial^2_{x} A_z + s_z \partial^2_{x} A_y \\ \dot{E}_y &= \int (s_z \partial_x f) dp ds \\ \dot{E}_z &= - \int (s_y \partial_x f) dp ds \end{aligned}\]

GEMPIC.strang_splitting! — Method

strang_splitting( h, dt, number_steps)

Strang splitting

time splitting object

time step

number of time steps