Particle-Mesh coupling

Spin version

GEMPIC.add_current_update_v! — Method

add_current_update_v!( j_dofs, p, 
                       position_old, position_new, 
                       marker_charge, qoverm, 
                       bfield_dofs, vi)

Add current for one particle and update v (according to $H_{p1}$ part in Hamiltonian splitting)

Read out particle position and velocity

Compute index_old, the index of the last DoF on the grid the

particle contributes to, and r_old, its position (normalized to cell size one).

GEMPIC.add_current_update_v_pp! — Method

add_current_update_v_pp!( j_dofs, p, position_old, position_new, 
                          marker_charge, qoverm, bfield_dofs, vi)

Add current for one particle and update v (according to H_p1 part in Hamiltonian splitting)

GEMPIC.update_jv! — Method

update_jv!(j_dofs, p, 
           lower, upper, index, marker_charge, 
           qoverm, sign, vi, bfield_dofs)

Helper function for add_current_update_v.

GEMPIC.update_jv_pp! — Method

update_jv_pp!( j_dofs, p, lower, upper, index, marker_charge, 
               qoverm, vi, bfield_dofs)

Helper function for add_current_update_v.