Documentation

GEMPIC.jl Documentation

GEMPIC.Maxwell1DFEM — Type.

maxwell_solver = MaxwellFEM1D( domain, ncells, degree )

1D Maxwell spline finite element solver on a periodic grid

Lx : length of Periodic domain

delta_x : cell size

n_dofs : number of cells (and grid points)

sdeg0 : spline degree 0-forms

sdeg1 : spline degree 1-forms

mass_0 : coefficients of 0-form mass matrix

mass_1 : coefficients of 1-form mass matrix

eig_mass0 : eigenvalues of circulant 0-form mass matrix

eig_mass1 : eigenvalues of circulant 1-form mass matrix

eigweakampere : eigenvalues of circulant update matrix for Ampere

eigweakpoisson : eigenvalues of circulant update matrix for Poisson

plan_fw : fft plan (forward)

plan_bw : fft plan (backward)

source

GEMPIC.ParticleMeshCoupling — Type.

Kernel smoother for 2d with splines of arbitrary degree placed on a uniform mesh. Spline with index i starts at point i

Value of grid spacing along both directions.

Definition of the domain: domain(1,1:2) = x1min, x1max

Number of particles of underlying PIC method (processor local)

Degree of smoothing kernel spline

Number of intervals where spline non zero (spline_degree + 1)

Scaling factor depending on whether Galerkin or collocation

Number of quadrature points

scratch data for spline evaluation

more scratch data for spline evaluation

quadrature weights and points

source

GEMPIC.compute_b_from_e — Method.

Compute Bz from Ey using strong 1D Faraday equation for spline coefficients

\[B_z^{new}(x_j) = B_z^{old}(x_j) - \frac{\Delta t}{\Delta x} (E_y(x_j) - E_y(x_{j-1})\]

source

GEMPIC.compute_e_from_b — Method.

compute Ey from Bz using weak Ampere formulation

source

GEMPIC.compute_e_from_j — Method.

Compute E_i from j_i integrated over the time interval using weak Ampere formulation

source

GEMPIC.compute_rhs_from_function — Method.

computerhsfromfunction(self, func, degree, coefsdofs)

Compute the FEM right-hand-side for a given function f and periodic splines of given degree.

Its components are $\int f N_i dx$ where $N_i$ is the B-spline starting at $x_i$.

source

GEMPIC.eval_uniform_periodic_spline_curve — Method.

evaluniformperiodicsplinecurve( degree, scoef )

Evaluate uniform periodic spline curve defined by coefficients scoef at knots (which are the grid points)

source

GEMPIC.l2norm_squared — Method.

Compute square of the L2norm

source

GEMPIC.l2projection — Method.

Compute the L2 projection of a given function f on periodic splines of given degree

source

GEMPIC.SplinePP — Type.

degree : degree of 1d spline

polycoeffs : polycoeffs[i,j] coefficient of x^{deg+1-j} for ith B-spline function size= (degree+1, degree+1)

polycoeffsfp : poly_coeffs[i,j] coefficient of x^{deg+1-j} for ith B-spline function size= (degree+1, degree+1)

ncells : number of gridcells

scratchb : scratch data for bto_pp-converting

scratchp : scratch data for bto_pp-converting

source

GEMPIC.b_to_pp — Method.

Convert 1d spline in B form to spline in pp form with periodic boundary conditions

source

GEMPIC.get_charge — Method.

Get charge of particle (q * particle_weight)

source

GEMPIC.get_mass — Method.

Get mass of particle (m * particle_weight)

source

GEMPIC.get_v — Method.

Get velocities

source

GEMPIC.get_weights — Method.

Get particle weights

source

GEMPIC.get_x — Method.

Get position

source

GEMPIC.horner_1d — Method.

Perform a 1d hornerschema on the pp_coeffs at index

source

GEMPIC.set_common_weight — Method.

Set the common weight

source

GEMPIC.set_v — Method.

Set velocity of particle @ i

source

GEMPIC.set_weights — Method.

Set weights of particle @ i

source

GEMPIC.set_x — Method.

Set position of particle @ i

source

GEMPIC.uniform_bsplines_eval_basis — Method.

uniform_bsplines_eval_basis( spline_degree, normalized_offset, bspl )

UNIFORM B-SPLINE FUNCTIONS

Evaluate all non vanishing uniform B-Splines in unit cell.

Returns an array with the values of the b-splines of the requested degree, evaluated at a given cell offset. The cell size is normalized between 0 and 1, thus the offset given must be a number between 0 and 1.

Output: bspl(1:d+1)= Bd(-(d+1)/2+d+x),...,Bd(-(d+1)/2+x) with d=splinedegree and x=normalizedoffset where Bd=B{d-1}*B0 and B0=1_[-1/2,1/2] and * is convolution the following code can be used for comparison with deboor

do i=-d,d+1
    t(i+d+1)=real(i,8)
end do
call bsplvb(t,d+1,1,normalized_offset,d+1,out)

We also have the property (from the symmetry of the B-spline) out(1:d+1)= Bd(-(d+1)/2+xx),...,Bd(-(d+1)/2+d+xx),..., where xx=1-normalized_offset

source