Hamiltonian splitting

GEMPIC.HamiltonianSplitting — Type

HamiltonianSplitting( maxwell_solver,
                      kernel_smoother_0, kernel_smoother_1,
                      particle_group, e_dofs, b_dofs)

Hamiltonian splitting type for Vlasov-Maxwell

Integral over the spline function on each interval (order p+1)

Integral over the spline function on each interval (order p)

e_dofs describing the two components of the electric field

b_dofs describing the magnetic field

j_dofs for kernel representation of current density.

GEMPIC.lie_splitting! — Method

lie_splitting( h, dt, number_steps)

Lie splitting

GEMPIC.lie_splitting_back! — Method

lie_splitting_back(h, dt, number_steps)

Lie splitting (oposite ordering)

GEMPIC.strang_splitting! — Method

strang_splitting( h, dt, number_steps)

Strang splitting

time splitting object

time step

number of time steps