Splines

Splines

GEMPIC.eval_uniform_periodic_spline_curve — Method.

eval_uniform_periodic_spline_curve( degree, scoef )

Evaluate uniform periodic spline curve defined by coefficients scoef at knots (which are the grid points)

source

GEMPIC.uniform_bsplines_eval_basis — Method.

uniform_bsplines_eval_basis( spline_degree, normalized_offset, bspl )

UNIFORM B-SPLINE FUNCTIONS

Evaluate all non vanishing uniform B-Splines in unit cell.

Returns an array with the values of the b-splines of the requested degree, evaluated at a given cell offset. The cell size is normalized between 0 and 1, thus the offset given must be a number between 0 and 1.

Output:

\[bspl(1:d+1)= B_d(-(d+1)/2+d+x),...,B_d(-(d+1)/2+x)\]

with $d$=spline_degree and $x$=normalized_offset where $B_d=B_{d-1}*B_0$ and $B_0=1_[-1/2,1/2]$ and * is convolution the following FORTRAN code can be used for comparison with deboor

do i=-d,d+1
    t(i+d+1)=real(i,8)
end do
call bsplvb(t,d+1,1,normalized_offset,d+1,out)

We also have the property (from the symmetry of the B-spline)

\[out[1:d+1]= B_d(-(d+1)/2+xx),...,B_d(-(d+1)/2+d+xx),..., \]

where $xx=1-$ normalized_offset

source