Mesh

GEMPIC.Mesh — Type

Mesh( xmin, xmax,nx )

Simple structure to store mesh data from 1 to 3 dimensions

GEMPIC.get_x — Method

get_x( mesh, i )

Get position

GEMPIC.get_cell_and_offset — Method

get_cell_and_offset( mesh, x )

Get cell and offset

We compute the cell indices where the particle is and its relative normalized position inside the cell