Mesh

GEMPIC.OneDGrid — Type

OneDGrid( xmin, xmax, nx )

Simple structure to store mesh data from 1 dimension

GEMPIC.ThreeDGrid — Type

ThreeDGrid( dimx, nx, dimy, ny, dimz, nz)

Generate a cartesians mesh on cube dimx x dimy x dimz with nx x ny x nz points

nx : indices are in [1:nx]

ny : indices are in [1:ny]

nz : indices are in [1:nz]

dimx = xmax - xmin

dimy = ymax - ymin

dimz = zmax - zmin

x, y, z : node positions

dx, dy, dz : step size

GEMPIC.TwoDGrid — Type

TwoDGrid( dimx, nx, dimy, ny)

Generate a cartesians mesh on rectangle dimxx dimy with nx x ny points

nx : indices are in [1:nx]

ny : indices are in [1:ny]

dimx = xmax - xmin

dimy = ymax - ymin

x, y : node positions

dx, dy : step size

GEMPIC.get_cell_and_offset — Method

get_cell_and_offset( mesh, x )

Get cell and offset

We compute the cell indices where the particle is and its relative normalized position inside the cell

GEMPIC.get_x — Method

get_x( mesh, i )

Get position