Particle group

Particle group

GEMPIC.ParticleGroup — Type.

ParticleGroup{D,V}( n_particles, 
                    charge, 
                    mass, 
                    n_weights)

n_particles : number of particles

charge : charge of one particle

mass : mass of one particle

n_weights : number of differents weights

source

GEMPIC.get_v — Method.

get_v( p, i )

Get velocity of ith particle of p

source

GEMPIC.get_x — Method.

get_x( p, i )

Get position of ith particle of p

source

GEMPIC.get_charge — Method.

get_charge( p, i; i_wi=1)

Get charge of ith particle of p (q * particle_weight)

source

GEMPIC.get_mass — Method.

get_mass( p, i; i_wi=1)

Get mass of ith particle of p (m * particle_weight)

source

GEMPIC.get_weights — Method.

get_weights( p, i)

Get ith particle weights of group p

source

GEMPIC.set_common_weight — Method.

set_common_weight( p, x )

Set the common weight

source

GEMPIC.set_v — Method.

set_v( p, i, v)

Set velocity of ith particle of p to v

source

GEMPIC.set_v — Method.

set_v( p, i, v)

Set velocity of ith particle of p to v

source

GEMPIC.set_weights — Method.

set_weights( p, i, w)

Set weights of ith particle of p to w

source

GEMPIC.set_weights — Method.

set_weights( p, i, w)

Set weights of particle @ i

source

GEMPIC.set_x — Method.

set_x( p, i, x )

Set position of ith particle of p to x

source

GEMPIC.set_x — Method.

set_x( p, i, x)

Set position of ith particle of p to x

source