Particle group

GEMPIC.ParticleGroup — Type

ParticleGroup{D,V}( n_particles, charge, mass, q, weights)

n_particles : number of particles

charge : charge of the particle species

mass : mass of the particle species

n_weights : number of differents weights

GEMPIC.get_charge — Method

get_charge( p, i; i_wi=1)

Get charge of ith particle of p (q * particle_weight)

GEMPIC.get_mass — Method

get_mass( p, i; i_wi=1)

Get mass of ith particle of p (m * particle_weight)

GEMPIC.get_spin — Method

get_spin( p, i, j)

Get the jth weight of the ith particle weights of group p

GEMPIC.get_v — Method

get_v( p, i )

Get velocity of ith particle of p

GEMPIC.get_weights — Method

get_weights( p, i)

Get ith particle weights of group p

GEMPIC.get_x — Method

get_x( p, i )

Get position of ith particle of p

GEMPIC.set_spin! — Method

set_spin!( p, i, j)

Set the jth weight of the ith particle weights of group p

GEMPIC.set_v! — Method

set_v!( p, i, v)

Set velocity of ith particle of p to v

GEMPIC.set_v! — Method

set_v!( p, i, v)

Set velocity of ith particle of p to v

GEMPIC.set_weights! — Method

set_weights!( p, i, w)

Set weights of particle @ i

GEMPIC.set_weights! — Method

set_weights!( p, i, w)

Set weights of ith particle of p to w

GEMPIC.set_x! — Method

set_x!( p, i, x)

Set position of ith particle of p to x

Note

if x is a scalar value, only the first x dimension will be set.

GEMPIC.set_x! — Method

set_x!( p, i, x )

Set position of ith particle of p to x