Particle-Mesh coupling

GEMPIC.ParticleMeshCoupling — Type

ParticleMeshCoupling( domain, n_grid, 
                      no_particles, spline_degree, 
                      smoothing_type )

Kernel smoother with splines of arbitrary degree placed on a uniform mesh. Spline with index i starts at point i

delta_x : Value of grid spacing along both directions.

domain : Definition of the domain: domain(1:2) = x1min, x1max

n_grid : Array containing number ofpoints along each direction

no_particles : Number of particles of underlying PIC method

spline_degree : Degree of smoothing kernel spline

n_span : Number of intervals where spline non zero (spline_degree + 1)

scaling : Scaling factor depending on whether :galerkin or :collocation

n_quad_points : Number of quadrature points

spline_val: scratch data for spline evaluation

spline_val_more : more scratch data for spline evaluation

quad_x, quad_w : quadrature weights and points

Note

Only 1D version is implemented for now

GEMPIC.evaluate — Method

evaluate(p, position, field_dofs)

Evaluate field at position

p : Kernel smoother object

position : Position of the particle

field_dofs : Coefficient vector for the field DoFs

field_value : Value(s) of the electric fields at given position

GEMPIC.add_charge! — Method

add_charge!( rho, p, position, marker_charge)

Add charge of one particle

p : kernel smoother object

position : Position of the particle

marker_charge : Particle weights time charge

rho_dofs : Coefficient vector of the charge distribution

GEMPIC.add_charge_pp! — Method

add_charge_pp!(rho_dofs, p, position, marker_charge)

Add charge of one particle

p : kernel smoother object

position : Position of the particle

marker_charge : Particle weights time charge

rho_dofs : Coefficient vector of the charge distribution

GEMPIC.add_current_update_v! — Method

add_current_update_v!( j_dofs, p, 
                       position_old, position_new, 
                       marker_charge, qoverm, 
                       bfield_dofs, vi)

Add current for one particle and update v (according to $H_{p1}$ part in Hamiltonian splitting)

Read out particle position and velocity

Compute index_old, the index of the last DoF on the grid the

particle contributes to, and r_old, its position (normalized to cell size one).

GEMPIC.add_current_update_v_pp! — Method

add_current_update_v_pp!( j_dofs, p, position_old, position_new, 
                          marker_charge, qoverm, bfield_dofs, vi)

Add current for one particle and update v (according to H_p1 part in Hamiltonian splitting)

GEMPIC.evaluate_pp — Method

evaluate_pp(p, position, field_dofs_pp)

Evaluate field at position using horner scheme

p : Kernel smoother object

position : Position of the particle

field_dofs_pp : Degrees of freedom in kernel representation.

field_value : Value(s) of the electric fields at given position

GEMPIC.update_jv! — Method

update_jv!(j_dofs, p, 
           lower, upper, index, marker_charge, 
           qoverm, sign, vi, bfield_dofs)

Helper function for add_current_update_v.

GEMPIC.update_jv_pp! — Method

update_jv_pp!( j_dofs, p, lower, upper, index, marker_charge, 
               qoverm, vi, bfield_dofs)

Helper function for add_current_update_v.